+++ src/login.html
@@ -131,6 +131,7 @@
                     admin_only_webcams: false,
                     allow_view_other_webcams: true,
                     allow_view_other_users_list: true,
+                    activate_waiting_room: true,
                     chat_features: {
                         allow_chat: true,
                         allow_file_upload: true,
